requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.2
asgiref==3.3.1
astroid==2.4.2
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
attrs==23.2.0
Babel==2.9.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==3.3.0
blinker==1.8.2
blis==0.7.11
braintree==4.7.0
Brotli==1.1.0
bs4==0.0.2
cachetools==4.2.2
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cryptography==43.0.1
cssselect==1.2.0
cycler==0.10.0
cymem==2.0.8
decorator==5.0.7
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
goto-statement==1.2
grpcio==1.64.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
httpx-socks==0.9.1
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
ipykernel==5.5.4
ipython==8.12.3
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.18.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.5
jsonschema==3.2.0
langcodes==3.4.0
language_data==1.2.0
lazy-object-proxy==1.4.3
libclang==18.1.1
lockfile==0.12.2
loguru==0.7.2
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mysql-connector-python==9.0.0
namex==0.0.8
nest-asyncio==1.5.1
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.1.0
openai==1.41.0
opencv-python==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandocfilters==1.4.3
pillow==10.4.0
preshed==3.0.9
protobuf==4.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.0.1
pymongo==4.8.0
pyobjc-core==10.3; sys_platform == 'darwin'
pyobjc-framework-Cocoa==10.3; sys_platform == 'darwin'
pyobjc-framework-Quartz==10.3; sys_platform == 'darwin'
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-socks==2.4.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.0
rich==13.7.1
rsa==4.7.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
selenium==4.23.1
sentence-transformers==3.0.1
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.1
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tika==2.6.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
--extra-index-url https://download.pytorch.org/whl/cpu
tqdm==4.66.5
traitlets==5.14.3
transformers==4.42.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.12.0
yarg==0.1.9
yarl==1.11.1
zipp==3.19.2
